+++ mysql/datadog_checks/mysql/config_models/defaults.py
@@ -24,7 +24,7 @@ def instance_custom_queries(field, value):
     return get_default_field_value(field, value)
 
 
-def instance_deep_database_monitoring(field, value):
+def instance_dbm(field, value):
     return False
 
 
@@ -48,11 +48,15 @@ def instance_min_collection_interval(field, value):
     return 15
 
 
+def instance_obfuscator_options(field, value):
+    return get_default_field_value(field, value)
+
+
 def instance_options(field, value):
     return get_default_field_value(field, value)
 
 
-def instance_pass_(field, value):
+def instance_password(field, value):
     return get_default_field_value(field, value)
 
 
@@ -61,6 +65,14 @@ def instance_port(field, value):
 
 
 def instance_queries(field, value):
+    return get_default_field_value(field, value)
+
+
+def instance_query_metrics(field, value):
+    return get_default_field_value(field, value)
+
+
+def instance_query_samples(field, value):
     return get_default_field_value(field, value)
 
 
@@ -76,10 +88,6 @@ def instance_ssl(field, value):
     return get_default_field_value(field, value)
 
 
-def instance_statement_samples(field, value):
-    return get_default_field_value(field, value)
-
-
 def instance_tags(field, value):
     return get_default_field_value(field, value)
 
@@ -88,5 +96,5 @@ def instance_use_global_custom_queries(field, value):
     return 'true'
 
 
-def instance_user(field, value):
+def instance_username(field, value):
     return 'datadog'